Use custom IP resolver for EC2 to get private IP addresses.
Using private IP addresses fixes an issue with how Hadoop determines what
interface to listen on. It also allows us to restrict communication to within
the security group, which is much more secure given that we still need to run on
a public subnet in order to give the nodes public internet access.