Allow services to record exceptions and include them in test summary
In some cases such as BackgroundThreadService, it is easy to lose the true cause of errors in the debug log if they don't
immediately kill the test. This allows these errors to be surfaced by allowing Services to optionally record error
information that will be aggregated at the end of a test and included in the test summary.