SSH command failure will either raise exception or log error
Current, when an SSH command fails, an exception will be raised only if `allow_fail` is false; if it is true, the result of any failed SSH command will be silently ignored.

This PR changes the behavior to log a debug message with the failure information when `allow_fail` is true. All other behavior is unchanged.