Add python3 support (#215)
This adds python3 to the standard unit test matrix and is verified to
work with the ducktape system tests on both py27 and py37. Most issues
were found easily with existing unit tests. System tests revealed a
couple of additional problems with str/bytes differences in py3 and
in Vagrant cluster code that is more difficult to unit test.